+++ packages/unlock-js/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - move `customHookService` to web3Service
 - add `getPasswordHookSigners` to web3Service
+- move `setPasswordHookSigner` to walletService
 
 # 0.33.4